# club_cms/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Club CMS API",
        "description": "Documentación de la API para el Club CMS",
        "contact": {
            "email": "soporte@clubcms.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost/api",
            "description": "Base URL para todos los endpoints"
        }
    ],
    "paths": {
...
        "/members": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Listar miembros",
                "description": "Obtiene una lista paginada de todos los miembros. Accesible para usuarios y administradores autenticados.",
                "operationId": "ffff12fc87590aad06f6e8282bd0831c",
                "responses": {
                    "200": {
                        "description": "Listado de miembros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "members": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/MemberResource"
                                            }
                                        },
                                        "page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total_pages": {
                                            "type": "integer",
                                            "example": 5
                                        },
                                        "total_members": {
                                            "type": "integer",
                                            "example": 50
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado - Token inválido o faltante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No hay miembros disponibles",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener los miembros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Members"
                ],
                "summary": "Crear un nuevo miembro (Solo Admins)",
                "description": "Crea un nuevo miembro en el sistema. Requiere rol de administrador.",
                "operationId": "3ba7591543d7d445e19be1c7fef65e9a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MemberRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Miembro creado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberCreateResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado - Token inválido o faltante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Miembro no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al crear el miembro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/members/{id}": {
            "get": {
                "tags": [
                    "Members"
                ],
                "summary": "Mostrar detalles de un miembro",
                "description": "Obtiene los detalles de un miembro específico. Accesible para usuarios y administradores autenticados.",
                "operationId": "1ec8ec404243666a7250dfdbc904733c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del miembro",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 42
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalle del miembro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "member": {
                                            "$ref": "#/components/schemas/MemberResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado - Token inválido o faltante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Miembro no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al obtener el miembro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Members"
                ],
                "summary": "Actualizar un miembro existente (Solo Admins)",
                "description": "Actualiza los datos de un miembro existente. Requiere rol de administrador.",
                "operationId": "b3343b4b4e177d072994080a4e19d263",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del miembro a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 42
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/MemberRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Miembro actualizado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/MemberUpdateResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado - Token inválido o faltante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Miembro no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al actualizar el miembro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Members"
                ],
                "summary": "Eliminar un miembro (Solo Admins)",
                "description": "Elimina un miembro del sistema. Requiere rol de administrador.",
                "operationId": "60ccb06330d81e74342a654c4683369d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del miembro a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 42
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Miembro eliminado correctamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "ok": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Miembro eliminado correctamente"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autorizado - Token inválido o faltante",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Miembro no encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error al eliminar el miembro",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "MemberRequest": {
                "required": [
                    "name",
                    "membership_date",
                    "status"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@email.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+34 123 456 789"
                    },
                    "membership_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "suspended"
                        ],
                        "example": "active"
                    }
                },
                "type": "object"
            },
            "MemberResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 42
                    },
                    "name": {
                        "type": "string",
                        "example": "Juan Pérez"
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "example": "juan@email.com"
                    },
                    "phone": {
                        "type": "string",
                        "example": "+34 123 456 789"
                    },
                    "membership_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-01-01"
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "suspended"
                        ],
                        "example": "active"
                    },
                    "has_system_access": {
                        "type": "boolean",
                        "example": false
                    },
                    "user": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 12
                            },
                            "name": {
                                "type": "string",
                                "example": "Juan Pérez"
                            },
                            "role": {
                                "type": "string",
                                "enum": [
                                    "user",
                                    "admin"
                                ],
                                "example": "user"
                            }
                        },
                        "type": "object",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "MemberCreateResponse": {
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": true
                    },
                    "message": {
                        "type": "string",
                        "example": "Miembro creado correctamente"
                    },
                    "member": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 32
                            },
                            "name": {
                                "type": "string",
                                "example": "Juan Pérez"
                            },
                            "membership_date": {
                                "type": "string",
                                "format": "date",
                                "example": "2023-01-01"
                            },
                            "status": {
                                "type": "string",
                                "enum": [
                                    "active",
                                    "inactive",
                                    "suspended"
                                ],
                                "example": "active"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "ErrorResponse": {
                "description": "Respuesta estándar para errores internos del servidor",
                "properties": {
                    "ok": {
                        "type": "boolean",
                        "example": false
                    },
                    "message": {
                        "description": "Mensaje de error seguro que no expone detalles técnicos",
                        "type": "string",
                        "example": "Error interno del servidor al procesar la solicitud."
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}